added a sound library changed some stuff in the components so they already have a render/update component so if i don't use it i don't have to add it to the file

+++ Minigin/Font.h
@@ -14,5 +14,5 @@ public:
 	Font& operator= (const Font&&) = delete;
 private:
 	_TTF_Font* m_Font;
-	unsigned int m_Size;
+	size_t m_Size;
 };
